third_party/WebKit/LayoutTests/inspector/console/console-trace-in-eval-expected.txt

CONSOLE MESSAGE: line 3: console.trace
Tests that when console.trace is called in eval'ed script ending with //# sourceURL=url it will dump a stack trace that will have the url as the script source. Bug 47252.

evalURL.js:3 console.trace
b @ evalURL.js:3
a @ evalURL.js:7
evalSource @ evalURL.js:10
(anonymous) @ evalURL.js:11
(anonymous) @ console-trace-in-eval.html:20
setTimeout (async)
doEvalSource @ console-trace-in-eval.html:19
(anonymous) @ VM:1
